# Add Generic Enum.TryFormat()