--- a/src/libraries/System.Private.Xml/src/System/Xml/Xsl/IlGen/OptimizerPatterns.cs
+++ b/src/libraries/System.Private.Xml/src/System/Xml/Xsl/IlGen/OptimizerPatterns.cs
@@ -242,7 +242,7 @@ namespace System.Xml.Xsl.IlGen
         /// </summary>
         public void AddPattern(OptimizerPatternName pattern)
         {
-            Debug.Assert(Enum.IsDefined(typeof(OptimizerPatternName), pattern));
+            Debug.Assert(Enum.IsDefined(pattern));
             Debug.Assert((int)pattern < 32);
             Debug.Assert(!_isReadOnly, "This OptimizerPatterns instance is read-only.");
             _patterns |= (1 << (int)pattern);
@@ -253,7 +253,7 @@ namespace System.Xml.Xsl.IlGen
         /// </summary>
         public bool MatchesPattern(OptimizerPatternName pattern)
         {
-            Debug.Assert(Enum.IsDefined(typeof(OptimizerPatternName), pattern));
+            Debug.Assert(Enum.IsDefined(pattern));
             return (_patterns & (1 << (int)pattern)) != 0;
         }