… src/libraries/System.Runtime/tests/System/EnumTests.cs
… src/libraries/System.ServiceProcess.ServiceController/src/System/ServiceProcess/ServiceController.cs