Partial implementation of #57881 

### Benchmarks

Code: https://gist.github.com/heathbm/59783b3d9e08d9556d7a86fa6ee2cd20

Before implementation:  

|      Method |     Mean |    Error |   StdDev |   Median |      Min |      Max |  Gen 0 | Allocated |
|------------ |---------:|---------:|---------:|---------:|---------:|---------:|-------:|----------:|
|     FormatG | 33.96 ns | 0.722 ns | 0.676 ns | 33.80 ns | 33.07 ns | 35.16 ns | 0.0019 |      24 B |
|     FormatD | 27.03 ns | 0.149 ns | 0.132 ns | 27.00 ns | 26.84 ns | 27.29 ns | 0.0019 |      24 B |
|     FormatX | 35.46 ns | 0.737 ns | 0.789 ns | 35.42 ns | 33.99 ns | 36.74 ns | 0.0050 |      64 B |
|     FormatF | 33.69 ns | 0.270 ns | 0.225 ns | 33.73 ns | 33.17 ns | 33.95 ns | 0.0018 |      24 B |

After implementation:  

|      Method |      Mean |     Error |    StdDev |    Median |       Min |      Max |  Gen 0 | Allocated |
|------------ |----------:|----------:|----------:|----------:|----------:|---------:|-------:|----------:|
|     FormatG | 32.983 ns | 0.2221 ns | 0.1969 ns | 32.980 ns | 32.588 ns | 33.40 ns | 0.0018 |      24 B |
|  TryFormatG | 14.332 ns | 0.1794 ns | 0.1678 ns | 14.412 ns | 13.943 ns | 14.51 ns |      - |         - |
|     FormatD | 26.751 ns | 0.8298 ns | 0.9556 ns | 26.348 ns | 25.166 ns | 28.47 ns | 0.0019 |      24 B |
| TryFormatD |  9.906 ns | 0.1853 ns | 0.1734 ns |  9.862 ns |  9.747 ns | 10.35 ns |      - |         - |
|     FormatX | 34.408 ns | 0.2204 ns | 0.1954 ns | 34.430 ns | 33.866 ns | 34.63 ns | 0.0050 |      64 B |
|  TryFormatX | 17.993 ns | 0.0465 ns | 0.0435 ns | 17.998 ns | 17.913 ns | 18.08 ns |      - |         - |
|     FormatF | 34.604 ns | 0.2278 ns | 0.2131 ns | 34.486 ns | 34.385 ns | 34.99 ns | 0.0018 |      24 B |
|  TryFormatF | 17.251 ns | 0.0471 ns | 0.0368 ns | 17.247 ns | 17.172 ns | 17.31 ns |      - |         - |

No visible regression.  

### Profiling:
Profiling code: https://gist.github.com/heathbm/079c4bea1dc541ffc12c91f283892875
Tested with:
- Visual Studio Performance Profiler (.NET Object Allocation Tracking)
- DotMemory (Full, Entire process tree)

Only allocations appear in the first run from:
- System.Enum.GetEnumInfo
- System.RuntimeType.GetTypeCodeImpl
I believe there is nothing I can do to prevent these allocations on the first run as they generate cache.

![alloc1](https://user-images.githubusercontent.com/13576170/177027167-99c8d868-f498-4dbf-a8a3-bf5420e0bfed.png)
![alloc2](https://user-images.githubusercontent.com/13576170/177027198-c66e11ff-9905-42c4-92b4-40aed96f5ad5.png)