@heathbm 2022-07-13
> The Enum tests are failing.

@jkotas I have removed a Debug.Assert that was unreliable since that assert can now deal with externally provide spans that can be larger than what was originally expected. I did not see any way to elegantly satisfy the assert without additional operations just to make the assert pass.