@jkotas 2022-09-06
It was addressed, thought there is still pretty significant code-bloat per instantiations that could be avoided by more significant refactoring.

Are there places in this repo where this method can be used?