@stephentoub 2022-09-27
> Are there places in this repo where this method can be used?

My expectation is we'll use it at a minimum in our interpolated string handlers, to avoid the boxing and string allocations that come from falling back to using ToString today.  We can start by calling this explicitly after doing some type testing (e.g. `typeof(T).IsEnum` when it's an intrinsic), though with the constraints on the public signature we'll probably need to move the logic out into an unconstrained helper.  Then that special-casing can be removed if/when Enum implements ISpanFormattable and the JIT special-cases it, though at that point I expect this implementation would be used by that ISpanFormattable implementation.