@jkotas 2022-11-08
> What kind of regressions, and do you have an example of where we'd expect this?

There seems to be some extra work done on paths through the existing APIs. We should run enum microbenchmarks to see whether it is going to show up.